Improve error handling in Org Notification Templates

+++ components/email-mgt/org.wso2.carbon.email.mgt/src/main/java/org/wso2/carbon/email/mgt/DBBasedEmailTemplateManager.java
@@ -31,12 +31,15 @@ import org.wso2.carbon.email.mgt.exceptions.I18nEmailMgtInternalException;
 import org.wso2.carbon.email.mgt.exceptions.I18nEmailMgtServerException;
 import org.wso2.carbon.email.mgt.model.EmailTemplate;
 import org.wso2.carbon.email.mgt.util.I18nEmailUtil;
+import org.wso2.carbon.identity.governance.IdentityMgtConstants;
 import org.wso2.carbon.identity.governance.exceptions.notiification.NotificationTemplateManagerClientException;
 import org.wso2.carbon.identity.governance.exceptions.notiification.NotificationTemplateManagerServerException;
 import org.wso2.carbon.identity.governance.model.NotificationTemplate;
+import org.wso2.carbon.identity.governance.service.notification.NotificationChannels;
 
 import java.util.List;
 
+import static org.wso2.carbon.email.mgt.constants.I18nMgtConstants.DEFAULT_EMAIL_LOCALE;
 import static org.wso2.carbon.email.mgt.constants.I18nMgtConstants.EMAIL_CHANNEL;
 import static org.wso2.carbon.email.mgt.util.I18nEmailUtil.*;
 
@@ -132,11 +135,31 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
     public void addEmailTemplate(EmailTemplate emailTemplate, String tenantDomain) throws I18nEmailMgtException {
 
         log.info("Test org addEmailTemplate(): " + emailTemplate.getTemplateDisplayName() + " locale: " + emailTemplate.getLocale());
+
+        NotificationTemplate notificationTemplate = buildNotificationTemplateFromEmailTemplate(emailTemplate);
         try {
-            NotificationTemplate notificationTemplate = buildNotificationTemplateFromEmailTemplate(emailTemplate);
+            //TODO: Check this can be moved to the API layer
+            validateNotificationTemplate(notificationTemplate);
+            I18nEmailUtil.getNormalizedName(notificationTemplate.getDisplayName());
+        } catch (NotificationTemplateManagerClientException e) {
+            throw new I18nEmailMgtClientException(e.getMessage(), e);
+        }
+
+        // TODO: Registry impl creates template type if not exists
+
+        if (isEmailTemplateExists(emailTemplate.getTemplateDisplayName(), emailTemplate.getLocale(), tenantDomain)) {
+            String locale = notificationTemplate.getLocale();
+            String scenarioName = notificationTemplate.getDisplayName();
+            String channel = notificationTemplate.getNotificationChannel();
+            // TODO: Check with the registry impl, it doesn't have a error here.
+            String message = String.format(I18nMgtConstants.ErrorMessages.ERROR_CODE_DUPLICATE_ORG_TEMPLATE.getMessage(), locale, scenarioName, channel, tenantDomain);
+            throw new I18nEmailMgtInternalException(I18nMgtConstants.ErrorCodes.EMAIL_TEMPLATE_ALREADY_EXISTS, message);
+        }
+
+        try {
             orgNotificationTemplateDAO.addNotificationTemplate(notificationTemplate, EMAIL_CHANNEL, getTenantId(tenantDomain));
-        } catch (Exception e) {
-            throw new I18nEmailMgtServerException("Error while adding the template", e);
+        } catch (NotificationTemplateManagerServerException e) {
+            throw new I18nEmailMgtServerException(e.getMessage(), e);
         }
     }
 
@@ -144,12 +167,49 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
     public EmailTemplate getEmailTemplate(String templateTypeDisplayName, String locale, String tenantDomain) throws I18nEmailMgtException {
 
         log.info("Test org getEmailTemplate(): " + templateTypeDisplayName + " locale: " + locale);
+
         try {
-            NotificationTemplate notificationTemplate = orgNotificationTemplateDAO.getNotificationTemplate(locale, templateTypeDisplayName, EMAIL_CHANNEL, getTenantId(tenantDomain));
-            return convertToEmailTemplate(notificationTemplate);
-        } catch (Exception e) {
-            throw new I18nEmailMgtServerException("Error while retrieving the template", e);
+            validateTemplateLocale(locale);
+            validateDisplayNameOfTemplateType(templateTypeDisplayName);
+        } catch (NotificationTemplateManagerClientException e) {
+            //TODO: ERROR_CODE_EMPTY_LOCALE & ERROR_CODE_EMPTY_TEMPLATE_NAME error codes are not handled in registry impl
+            throw new I18nEmailMgtClientException(e.getMessage(), e);
         }
+
+        String notificationChannel = EMAIL_CHANNEL;
+        NotificationTemplate notificationTemplate;
+        try {
+            notificationTemplate = orgNotificationTemplateDAO.getNotificationTemplate(locale, templateTypeDisplayName, notificationChannel, getTenantId(tenantDomain));
+        } catch (NotificationTemplateManagerServerException e) {
+            throw new I18nEmailMgtServerException(e.getMessage(), e);
+        }
+
+        // TODO: Check why its only SMS here in registry impl
+        // Handle not having the requested SMS template type in required locale for this tenantDomain.
+        if (notificationTemplate == null) {
+            String defaultLocale = getDefaultNotificationLocale(notificationChannel);
+            if (StringUtils.equalsIgnoreCase(defaultLocale, locale)) {
+                // Template is not available in the default locale. Therefore, breaking the flow at the consuming side to avoid NPE.
+                String error = String.format(IdentityMgtConstants.ErrorMessages.ERROR_CODE_NO_TEMPLATE_FOUND.getMessage(), templateTypeDisplayName, locale, tenantDomain);
+                throw new I18nEmailMgtInternalException(I18nMgtConstants.ErrorCodes.EMAIL_TEMPLATE_TYPE_NODE_FOUND, error);
+            } else {
+                if (log.isDebugEnabled()) {
+                    String message = String.format("'%s' template in '%s' locale was not found in '%s' tenant. Trying to return the template in default locale : '%s'", templateTypeDisplayName, locale, tenantDomain, DEFAULT_EMAIL_LOCALE);
+                    log.debug(message);
+                }
+                // Try to get the template type in default locale.
+                return getEmailTemplate(templateTypeDisplayName, defaultLocale, tenantDomain);
+            }
+        }
+
+        // TODO: Remove following as its not needed when retrieving, but registry impl has it
+        if (StringUtils.isBlank(notificationTemplate.getBody())) {
+            String error = String.format(IdentityMgtConstants.ErrorMessages.
+                    ERROR_CODE_NO_CONTENT_IN_TEMPLATE.getMessage(), templateTypeDisplayName, locale);
+            throw new I18nEmailMgtClientException(IdentityMgtConstants.ErrorMessages.
+                    ERROR_CODE_NO_CONTENT_IN_TEMPLATE.getCode(), error);
+        }
+        return buildEmailTemplate(notificationTemplate);
     }
 
     @Override
@@ -157,13 +217,21 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
             throws I18nEmailMgtException {
 
         log.info("Test org isEmailTemplateExists(): " + templateTypeDisplayName + " locale: " + locale);
+
+//        try {
+        //TODO: Check this can be moved to the API layer
+//            EmailTemplateManagerImpl.validateDisplayNameOfTemplateType(templateTypeDisplayName);
+//        } catch (NotificationTemplateManagerClientException e) {
+//            throw new I18nEmailMgtClientException(e.getMessage(), e);
+//        }
+        I18nEmailUtil.getNormalizedName(templateTypeDisplayName);
+
         try {
             NotificationTemplate notificationTemplate = orgNotificationTemplateDAO.getNotificationTemplate(locale, templateTypeDisplayName, EMAIL_CHANNEL, getTenantId(tenantDomain));
             return notificationTemplate != null;
-        } catch (Exception e) {
-            throw new I18nEmailMgtServerException("Error while checking the existence of the template", e);
+        } catch (NotificationTemplateManagerServerException e) {
+            throw new I18nEmailMgtServerException(e.getMessage(), e);
         }
-
     }
 
     @Override
@@ -171,10 +239,19 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
             throws I18nEmailMgtException {
 
         log.info("Test org getEmailTemplate(): " + templateDisplayName);
+
+        try {
+            //TODO: Check this can be moved to the API layer
+            validateDisplayNameOfTemplateType(templateDisplayName);
+            I18nEmailUtil.getNormalizedName(templateDisplayName);
+        } catch (NotificationTemplateManagerClientException e) {
+            throw new I18nEmailMgtClientException(e.getMessage(), e);
+        }
+
         try {
             List<NotificationTemplate> notificationTemplates = orgNotificationTemplateDAO.listNotificationTemplates(templateDisplayName, EMAIL_CHANNEL, getTenantId(tenantDomain));
             return convertToEmailTemplates(notificationTemplates);
-        } catch (Exception e) {
+        } catch (NotificationTemplateManagerServerException e) {
             throw new I18nEmailMgtServerException("Error while listing the templates", e);
         }
     }
@@ -183,10 +260,21 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
     public void deleteEmailTemplate(String templateTypeDisplayName, String locale, String tenantDomain) throws I18nEmailMgtException {
 
         log.info("Test org deleteEmailTemplate(): " + templateTypeDisplayName + " locale: " + locale);
+
+        //TODO: Check this can be moved to the API layer
+        // Validate the name and locale code.
+        if (StringUtils.isBlank(templateTypeDisplayName)) {
+            throw new I18nEmailMgtClientException("Cannot Delete template. Email displayName cannot be null.");
+        }
+        if (StringUtils.isBlank(locale)) {
+            throw new I18nEmailMgtClientException("Cannot Delete template. Email locale cannot be null.");
+        }
+        I18nEmailUtil.getNormalizedName(templateTypeDisplayName);
+
         try {
             orgNotificationTemplateDAO.removeNotificationTemplate(locale, templateTypeDisplayName, EMAIL_CHANNEL, getTenantId(tenantDomain));
-        } catch (Exception e) {
-            throw new I18nEmailMgtServerException("Error while deleting the template", e);
+        } catch (NotificationTemplateManagerServerException e) {
+            throw new I18nEmailMgtServerException(e.getMessage(), e);
         }
     }
 
@@ -194,10 +282,20 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
     public void deleteEmailTemplates(String templateDisplayName, String tenantDomain) throws I18nEmailMgtException {
 
         log.info("Test org deleteEmailTemplates(): " + templateDisplayName);
+
+        try {
+            //TODO: Check this can be moved to the API layer
+            validateDisplayNameOfTemplateType(templateDisplayName);
+            I18nEmailUtil.getNormalizedName(templateDisplayName);
+        } catch (NotificationTemplateManagerClientException e) {
+            //TODO: error handling is different from the registry based impl
+            throw new I18nEmailMgtClientException(e.getMessage(), e);
+        }
+
         try {
             orgNotificationTemplateDAO.removeNotificationTemplates(templateDisplayName, EMAIL_CHANNEL, getTenantId(tenantDomain));
-        } catch (Exception e) {
-            throw new I18nEmailMgtServerException("Error while deleting the template", e);
+        } catch (NotificationTemplateManagerServerException e) {
+            throw new I18nEmailMgtServerException(e.getMessage(), e);
         }
     }
 
@@ -219,7 +317,7 @@ public class DBBasedEmailTemplateManager implements EmailTemplateManager {
         log.info("Test app getEmailTemplate(): " + templateDisplayName);
         try {
             NotificationTemplate notificationTemplate = appNotificationTemplateDAO.getNotificationTemplate(locale, templateDisplayName, EMAIL_CHANNEL, applicationUuid, getTenantId(tenantDomain));
-            return convertToEmailTemplate(notificationTemplate);
+            return buildEmailTemplate(notificationTemplate);
         } catch (Exception e) {
             throw new I18nEmailMgtServerException("Error while retrieving the app template", e);
         }
